try to draw triangle

+++ Music_Tuner_2/GameManager.h
@@ -33,7 +33,7 @@ private:
 	//std::unique_ptr<IGameState> m_currentState;
 	bool m_isGameFinished; // ★追加: ゲーム終了フラグ
 	InputState m_inputState;
-	float m_prevTime = 0;
+	LONGLONG m_prevTime = 0;
 
 	void UpdateInputState();
